refactor(runtime): separate transaction envelope context and health
