+++ src/runtime/src/runtime/program_transaction.rs
@@ -5,7 +5,6 @@
 //! live-runtime, and context state.
 
 use super::*;
-use crate::{AccessSet, RuntimeAuthorityScope};
 #[cfg(feature = "invariant_define")]
 use mech_program::{
   IntegrityConstraintFailureReason, IntegrityConstraintViolationSet,
@@ -68,241 +67,12 @@ pub(super) fn integrity_failure_audit(
   })
 }
 
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(super) enum RuntimeExecutionTransactionMode {
-  Explicit,
-  ImplicitModuleOperation,
-  ImplicitProgramOperation,
-  ImplicitReactiveTurn,
-  ImplicitResourceOperation,
-}
-
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(super) enum RuntimeExecutionTransactionState {
-  Active,
-  Committing,
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub(super) struct RuntimeTransactionContextIdentity {
-  runtime: RuntimeId,
-  subject: String,
-  task: Option<TaskId>,
-  actor: Option<ActorId>,
-  actor_message: Option<MessageRecord>,
-  actor_state: Option<ObjectId>,
-}
-
-impl RuntimeTransactionContextIdentity {
-  pub(super) fn capture(context: &RuntimeContext) -> Self {
-    Self {
-      runtime: context.runtime,
-      subject: context.subject.clone(),
-      task: context.task,
-      actor: context.actor,
-      actor_message: context.actor_message.clone(),
-      actor_state: context.actor_state,
-    }
-  }
-
-  pub(super) fn mismatch_reason(&self, context: &RuntimeContext) -> Option<String> {
-    if self.runtime != context.runtime {
-      return Some(format!(
-        "runtime changed from {} to {}",
-        self.runtime,
-        context.runtime,
-      ));
-    }
-    if self.subject != context.subject {
-      return Some(format!(
-        "subject changed from `{}` to `{}`",
-        self.subject,
-        context.subject,
-      ));
-    }
-    if self.task != context.task {
-      return Some(format!(
-        "task changed from {:?} to {:?}",
-        self.task,
-        context.task,
-      ));
-    }
-    if self.actor != context.actor {
-      return Some(format!(
-        "actor changed from {:?} to {:?}",
-        self.actor,
-        context.actor,
-      ));
-    }
-    if self.actor_message != context.actor_message {
-      return Some("actor message changed".to_string());
-    }
-    if self.actor_state != context.actor_state {
-      return Some(format!(
-        "actor state changed from {:?} to {:?}",
-        self.actor_state,
-        context.actor_state,
-      ));
-    }
-    None
-  }
-
-  pub(super) fn set_actor_state(&mut self, state: ObjectId) {
-    self.actor_state = Some(state);
-  }
-
-  pub(super) fn bind_actor_turn(&mut self, turn: &ActorTurn) {
-    self.subject = turn.subject.clone();
-    self.actor = Some(turn.actor);
-    self.actor_message = Some(turn.message.clone());
-    self.actor_state = turn.state;
-  }
-}
-
-#[derive(Clone)]
-pub(super) struct RuntimeContextCheckpoint {
-  runtime: RuntimeId,
-  subject: String,
-  task: Option<TaskId>,
-  actor: Option<ActorId>,
-  access: AccessSet,
-  module_version: Option<ModuleVersionId>,
-  transaction: Option<TransactionId>,
-  authority: RuntimeAuthorityScope,
-  budget: ResourceBudget,
-  events: Vec<RuntimeEvent>,
-  actor_message: Option<MessageRecord>,
-  actor_state: Option<ObjectId>,
-}
-
-impl RuntimeContextCheckpoint {
-  pub(super) fn capture(context: &RuntimeContext) -> Self {
-    Self {
-      runtime: context.runtime,
-      subject: context.subject.clone(),
-      task: context.task,
-      actor: context.actor,
-      access: context.access.clone(),
-      module_version: context.module_version,
-      transaction: context.transaction,
-      authority: context.authority.clone(),
-      budget: context.budget.clone(),
-      events: context.events.clone(),
-      actor_message: context.actor_message.clone(),
-      actor_state: context.actor_state,
-    }
-  }
-
-  pub(super) fn restore_preserving_consumption(&self, context: &mut RuntimeContext) {
-    let used_steps = context.budget.used_steps.max(self.budget.used_steps);
-    let used_bytes = context.budget.used_bytes.max(self.budget.used_bytes);
-    let used_items = context.budget.used_items.max(self.budget.used_items);
-    let used_messages = context.budget.used_messages.max(self.budget.used_messages);
-
-    context.runtime = self.runtime;
-    context.subject = self.subject.clone();
-    context.task = self.task;
-    context.actor = self.actor;
-    context.access = self.access.clone();
-    context.module_version = self.module_version;
-    context.transaction = self.transaction;
-    context.authority = self.authority.clone();
-    context.budget = ResourceBudget {
-      max_steps: self.budget.max_steps,
-      used_steps,
-      max_bytes: self.budget.max_bytes,
-      used_bytes,
-      max_items: self.budget.max_items,
-      used_items,
-      max_messages: self.budget.max_messages,
-      used_messages,
-    };
-    context.events = self.events.clone();
-    context.actor_message = self.actor_message.clone();
-    context.actor_state = self.actor_state;
-  }
-
-  pub(super) fn access_delta(&self, context: &RuntimeContext) -> AccessSet {
-    AccessSet {
-      reads: context
-        .access
-        .reads
-        .iter()
-        .copied()
-        .filter(|object| !self.access.reads.contains(object))
-        .collect(),
-      writes: context
-        .access
-        .writes
-        .iter()
-        .copied()
-        .filter(|object| !self.access.writes.contains(object))
-        .collect(),
-    }
-  }
-}
-
-#[derive(Clone)]
-pub(super) struct RuntimeProgramBaseline {
-  pub(super) program: MechProgramCheckpoint,
-  pub(super) live: RuntimeLiveStateSnapshot,
-}
-
 pub(super) enum RuntimeProgramOwnershipAcquisition {
   Existing,
   NewlyAcquired {
     program: MechProgramCheckpoint,
     live: RuntimeLiveStateSnapshot,
   },
-}
-
-pub(super) struct RuntimeExecutionTransaction {
-  pub(super) store: RuntimeTransaction,
-  pub(super) modules: RuntimeModuleJournal,
-  pub(super) mode: RuntimeExecutionTransactionMode,
-  pub(super) context_identity: RuntimeTransactionContextIdentity,
-  pub(super) context_baseline: RuntimeContextCheckpoint,
-  pub(super) program: Option<RuntimeProgramBaseline>,
-  pub(super) effects: RuntimeEffectJournal,
-  pub(super) capabilities: RuntimeCapabilityOverlay,
-  pub(super) state: RuntimeExecutionTransactionState,
-}
-
-impl RuntimeExecutionTransaction {
-  pub(super) fn new(
-    store: RuntimeTransaction,
-    mode: RuntimeExecutionTransactionMode,
-    context_identity: RuntimeTransactionContextIdentity,
-    context_baseline: RuntimeContextCheckpoint,
-  ) -> Self {
-    Self {
-      store,
-      modules: RuntimeModuleJournal::new(),
-      mode,
-      context_identity,
-      context_baseline,
-      program: None,
-      effects: RuntimeEffectJournal::new(),
-      capabilities: RuntimeCapabilityOverlay::default(),
-      state: RuntimeExecutionTransactionState::Active,
-    }
-  }
-}
-
-#[derive(Clone)]
-pub(super) struct RuntimeOperationSavepoint {
-  pub(super) store: RuntimeTransaction,
-  pub(super) module_mark: usize,
-  pub(super) effect_mark: usize,
-  pub(super) capability_mark: usize,
-  pub(super) context: RuntimeContextCheckpoint,
-}
-
-#[derive(Clone)]
-pub(super) struct RuntimeProgramOperationSavepoint {
-  pub(super) program: MechProgramCheckpoint,
-  pub(super) live: RuntimeLiveStateSnapshot,
-  pub(super) runtime: RuntimeOperationSavepoint,
 }
 
 #[derive(Clone, Copy, Debug)]
@@ -350,20 +120,6 @@ pub(super) fn reset_runtime_program_checkpoint_count() {
 #[cfg(test)]
 pub(super) fn runtime_program_checkpoint_count() -> usize {
   RUNTIME_PROGRAM_CHECKPOINT_COUNT.with(std::cell::Cell::get)
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum RuntimeHealth {
-  Healthy,
-  Poisoned(RuntimePoisonRecord),
-}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub struct RuntimePoisonRecord {
-  pub operation: String,
-  pub transaction_id: Option<TransactionId>,
-  pub original_error: String,
-  pub rollback_failures: Vec<String>,
 }
 
 impl MechRuntime {
