+++ src/runtime/src/runtime/transaction/savepoint.rs
@@ -1,0 +1,20 @@
+use super::RuntimeContextCheckpoint;
+use crate::runtime::RuntimeLiveStateSnapshot;
+use crate::RuntimeTransaction;
+use mech_program::MechProgramCheckpoint;
+
+#[derive(Clone)]
+pub(in crate::runtime) struct RuntimeOperationSavepoint {
+  pub(in crate::runtime) store: RuntimeTransaction,
+  pub(in crate::runtime) module_mark: usize,
+  pub(in crate::runtime) effect_mark: usize,
+  pub(in crate::runtime) capability_mark: usize,
+  pub(in crate::runtime) context: RuntimeContextCheckpoint,
+}
+
+#[derive(Clone)]
+pub(in crate::runtime) struct RuntimeProgramOperationSavepoint {
+  pub(in crate::runtime) program: MechProgramCheckpoint,
+  pub(in crate::runtime) live: RuntimeLiveStateSnapshot,
+  pub(in crate::runtime) runtime: RuntimeOperationSavepoint,
+}
